Revise Friendship destroy method

+++ app/controllers/friendships_controller.rb
@@ -26,8 +26,10 @@ class FriendshipsController < ApplicationController
 
   def destroy
     friendship = Friendship.find_by(friend_id: params[:friend_id], user: current_user)
-    if friendship
+    friendship_second = Friendship.find_by(user_id: params[:friend_id], friend_id: current_user.id)
+    if friendship && friendship_second
       friendship.destroy
+      friendship_second.destroy
       redirect_to users_path, notice: 'You recalled friend request.'
     else
       redirect_to users_path, alert: 'You cannot recall friend request that you recalled already.'
